+++ src/host/get_z.c
@@ -2,20 +2,20 @@
 
 // samples -> get_z -> preactivation
 
-double *get_z(NETWORK *n, int layer_index, double *sample)
+float *get_z(NETWORK *n, int layer_index, float *sample)
 {
     LAYER *current_layer = n->l+layer_index;
     int z_neuroncount = current_layer->num_neurons;
     int z_weightcount = current_layer->n->num_weights;
     int is_first_layer = layer_index == 0;
 
-    double *z = (double *) malloc (sizeof(double)* z_neuroncount);
+    float *z = (float *) malloc (sizeof(float)* z_neuroncount);
     if(!z) {
         fprintf(stderr, "Error 10005\n");
         return NULL;
     }
 
-    double *z_prev = is_first_layer ? sample : get_y(n, layer_index-1, sample);
+    float *z_prev = is_first_layer ? sample : get_y(n, layer_index-1, sample);
 
     for(size_t i=0; i<z_neuroncount; ++i) {
         z[i] = 0;
