+++ src/host/sse.c
@@ -1,12 +1,12 @@
 #include "mlp.h"
 
-double sse(double *real, double *ideal, int length)
+float sse(float *real, float *ideal, int length)
 {
-    double sse = 0.0;   // Sum of squared errors
+    float sse = 0.0;   // Sum of squared errors
 
     for(size_t i=0; i<length; ++i) {
-        double raw_error = ideal[i]-real[i];
-        sse += pow(raw_error, 2);
+        float raw_error = ideal[i]-real[i];
+        sse += powf(raw_error, 2);
     }
 
     return sse;
